Move CPUReproTests to its own file
[ghstack-poisoned]
